Update migrate-from-classic-mode-to-claims-based-authentication-in-sharepoint-2013.md
Fixed links. Fixed formatting. Changed certain SharePoint 2016 references to SharePoint 2013. Changed the format of converting from a SharePoint 2010 Classic to SharePoint 2013 claims -- start out with a SharePoint 2013 claims-enabled web app and perform the conversion after mounting the databases.

Fixed the `Convert-SPWebApplication` cmdlet to specify that a `-From Legacy` is required as this was added in 2014 and is required in currently supported versions of SharePoint 2013.