Switch health rule article to use PowerShell
This is a new SharePoint 2019 article for the SMTP authentication health rule check.  It provides instructions to resolve the issue based on the stsadm.exe command line tool.  Since we now have a PowerShell cmdlet that does the same thing as that stsadm command, we should update our instructions to use that PowerShell cmdlet.