Provided an option to prevent any page modification
To me the article was confusing as I was looking for a way to prevent members to create any kind of pages. I found out that fiddling with the "Site Pages" feature only disables creation of modern pages, but still allows classic pages to be created. I don't really see in which scenario this would be useful. To provide people in the same scenario as me, so wanting to disallow any kind of page creation of modification, at least a hint towards how to accomplish this, I've added a hint towards how to do that to the notice box.