Update restrict-or-enable-access-to-a-service-application.md
Fix for #192 - removed reference instead of adding a section. The target audience for this specific article will be those who should have familiarity with PowerShell. We also don't outline the steps on how to launch/use PowerShell in other SharePoint articles of a similar nature.

Fixed some major formatting issues.