Updated information on using OneDrive with SharePoint 2019
The article still mentioned that the insider ring was required to use OneDrive with SharePoint 2019. This is no longer the case. The current production release will work with it as well. Furthermore it was a bit vague on which Group Policy to configure to get it to work and how to ensure you have the proper group policy. It also didn't make mention of the underlying registry keys which could be used as an alternative for configuring OneDrive to work with SharePoint 2019.