Updated companies and content class rows in table
As per Samuel Shelton (eDiscovery Dev Lead) you can't use eDiscovery to search companies and contentclass properties. The companies property is used for People search, not eDiscovery. It looks like contentclass is video related? So for both of these properties I changed the value in the "Queryable" column to "No". And for the companies property I added the value "People:companies" to the Mapped crawled properties column.