Update use-group-policy.md
"users who are signed in on a PC with the primary Windows account (the account used to join the PC to the domain) can set up the sync client without entering their account credentials. " 

This implies that only the account that was used to join the PC to the domain will get SilentAccount. However, based on initial testing, this does not appear to be case and multiple domain users, even though they didn't join the PC to the on premises AD domain (in the cases of Hybrid Azure AD Join), SilentAccount worked for them.