Multiple organization asset libraries on same site collection
Added notice on that you can add multiple document libraries as being organizational asset libraries, but all of these MUST reside on the same site collection or else an error will be thrown stating: "This library is not on the site that contains other organization asset libraries.". Good to point this out in the documentation.