Bold Option 1: Option 2: and NOTE:
Could we bold Option 1: , bold Option 2: and bold the NOTE: so it's more obvious to the reader? Don't need the entire sentence bolded, but would like to see those sections highlighted as go to points in the document.  Not sure how to make just a word or phrase bold in the editor vs. an entire phrase. thx.