Update software-boundaries-and-limits-0.md
This article contains misleading information.

The biggest discrepancy here is the fact that one time we recommend to stay below 250,000 web sites per site collection and further down we recommend to stay below 2000.
The next is that there is a lot of ambiguity between site, webs and subsites in this paragraph. We only have web sites. If we would like to talk about web sites that are inside other web sites we can say sub-web site.