[DO NOT MERGE] SEO edits for SharePoint file collaboration
Edited title to remove "- SharePoint." System automatically adds "- SharePoint in Microsoft 365" to the title.

Edited description to add keywords.

Per the SEO expert recommendation, expanded the "Manage access" heading to include the keyword "guest users." Added to this heading as it seems to be the only section that already mentions guest users multiple times.