[DO NOT MERGE] SEO edits for Group Policy to control OneDrive sync
Edited metadata title to use configure instead of control.

Edited description to add "configure".

Added an intro sentence to the Computer Configuration policies to insert the high-value keywords "disable OneDrive Group Policy Object". Per the SEO expert, this is a very high-hit phrase. The individual settings seem to indicate enabling and disabling, but I'm not sure if these are considered GPOs. If the sentence is technically inaccurate, I will remove it.