[DO NOT MERGE] SEO edits for SPMT supported features
Edited description per the SEO expert's recommendation for clarity and keywords.

There is a recommendation to add the keyword "migration tool limitations". I think that would add a negative connotation to the article, "what the SPMT can't do". This is a high-traffic keyword not addressed by other articles, so I could find a way to add it if the author's okay with it. But it would likely need to be in it's own heading that specifically called out the limitations, for which I'd need some source.