Add wording to line 38
The first section of the article says this:
"By default, script is allowed on sites that admins create. It is not allowed on OneDrive, on sites users create themselves, and on the root site for your organization."

That is true, but by default, it is also not allowed on Communication and Modern Team sites.  The DenyAddAndCustomizePages property is set to Enabled on those sites by default.  This omission has led to customers opening support cases.

I believe the wording should be updated to something like this:
"By default, script is allowed on most sites that admins create. It is not allowed on OneDrive, on sites users create themselves, on modern Team and Communications sites, and on the root site for your organization."