Adding a new setting to disable Setup tutorial
ALSO is it possible to alphabetize this list, similar to the one that we have for the Windows policy list.  The order of this list seems Wild, I dont know markdown super well so I'm assuming someone can do it faster than me.  I'd like to alphabetize the list based on the Settings column.