Clarified that locations need to be explicitly allowed in at least one DLP policy for sharing to work.
The requirement to create a DLP policy with all locations and all users was opaque and many customers were asking for clarification. I added an explanation about the source of this requirement, and detailed that the option to create a DLP policy with all sites and all users is only required if you don't want to require that locations are explicitly authorized in a DLP policy for them to be shareable.