Decreased indentation of flush-left images
In Markdown, image code should be aligned with the preceding text to which it is meant to be aligned—flush left for regular body paragraphs, indented for content in list items. I've removed the extra space here, because the image code is on the threshold of having too many spaces before it. When there's too much space, the code is rendered as a code block, not as an image.