Made several changes. Other changes suggestion below.
1. When you run `prerequisiteinstaller.exe` at a command prompt, you might be asked to restart the server one or more times during the installation process. After restarting, you should continue the prerequisite installation by running `prerequisiteinstaller.exe` with the /continue option.

- The above system can be added as a Note.
_________________________________________________
/? This displays command-line options.

/continue This is used to tell the installer that it is continuing from being restarted.

/unattended This indicates no user interaction.

The above content can be better represented in a table.

________________________________________________________________
"file" - The angular brackets were preventing the word to appear in the published page. I think code tag should be used for it. For now i have put it within quotes. There are multiple instances of this.
___________________________________

/quiet - in code tag.
______________________________________

o configure by using Server Manager, see Install or Uninstall Roles, Role Services, or Features - Is this suppose to be a link ? - Also end of sentence punctuation missing.
________________________________________________________

-Source D:\sources\sxs - code tag