Update request for assigning SharePoint license
OneDrive admin center settings are coming to the SharePoint admin center. 
https://www.microsoft.com/ja-jp/microsoft-365/roadmap?filters=&searchterms=65911
MC240393

Internal Wiki says :
Q. Does the customer who has OneDrive for Business license only need to be assigned a SPO license to access SharePoint admin center with this convergence?
A. Nothing is changing from the licensing and admin access permissions side with this transition.