Correct indentation & numbering of subordinate procedure
"To create a connection to a directory service for import" has appeared to have a 17- or 18-step procedure (depending on whether the unindented note in Step 17 is present) for quite some time. However, step 7 indicates that everything that follows is a procedure to complete that step. As such, these steps should be indented and use alphabetic enumeration.