Corrected indentation and numbering of subordinate procedure
The procedure under "To map user profile properties". In the Markdown source, this is clearly meant to have 3 subordinate steps under Step 6, but for quite some time, this section has been rendered as an 11-step procedure with the last step referring to procedure steps that are incorrectly numbered in the rendered page.