SPFx is not supposed to be used for branding
The only thing supported by SPFx regarding branding is header/footer customizations.
If you try to change theme elements, you will have a flashing experience at the start of the page load where you see the OOB page, and then the SPFx stuff is added. This is not avoidable and is why we have no support for it. You might want to add more about the supported ways, here are some related links-
https://docs.microsoft.com/en-us/sharepoint/dev/declarative-customization/site-theming/sharepoint-site-theming-overview
https://docs.microsoft.com/en-us/sharepoint/dev/transform/modernize-branding