Add note for top queries requirement
Would like to add a Note callout for the top queries requirement that consumes the following sentences. 

A query is added to this report when it's searched at least three times with a click on a result. Use this report to understand what types of information your users are searching for.