clarifying settings hierarchy
i've attempted to clarify that there is a master switch that controls all the other switches. in addition, i added a row for Teams Loop components with 'n/a' to clarify that there isn't currently available, a switch that only impacts Teams Loop components, that the master switch must be used. we know this is a customer dissat issue, and we're working on a new solution for settings and we'll need to update this article once coded.