prepping for IsLoopEnabled
at the end of this month, a new (anxiously awaited) setting will become available: IsLoopEnabled. this allows a tenant to disable Loop and Whiteboard independently, which was previously not possible. this update also moves IsFluidEnabled down to the bottom of the table in preparation for it being deprecated later this year.