Update subsite verbiage
Communication sites do not support the subsite feature at all. Team sites do. The idea is to state the lack of support for the communication site first before the reader proceeds with reading. The original wording implies that a user can create a communication site and then have team sites as web pages which is not supported.