Added note about navigation item limitation
The development team recently introduced a new limitation to the number of navigation items that can be contained on any single node in the navigation. This is for all navigation types.

Added the following note at the beginning of the article "All navigation types (site, hub, global, and footer) are limited to 500 child links at each level. Additional nodes added after the limit has been reached will receive an error."