create sensitive-by-default.md
Line 27 line: 
Instead of turning off external sharing entirely, you can mark the files in your organization as sensitive by default. This blocks guest access to new content until it has been scanned for sensitive content and DLP policies that include content-based conditions are applied. Guests are notified that the file is being scanned if they attempt to access it during this time.

This line needs to be changed because guests are unable to access the files are ARE NOT notified that the file is being scanned. Instead, they are receiving a generic error that causes confusion and creates unnecessary calls to support. A customer recently reported this in a support ticket and I was also able to reproduce this in my environment.
The errors were different depending on if they were accessing the file in the Web apps in the browser or the integrated apps within Teams.

Please make revisions to this article to set proper expectations on what the experience will be like.

I am attaching sample errors the customer was receiving. Nothing to indicate files are being scanned