Add Word to the Cloud Policy section
Word will be launching to First Release in a week or so, proactively updating the documentation to include this. I've also updated the top paragraph a bit to make it a little clearer that there are two types of settings and you have to think about both of them right now.