Clarify SharePoint AMSI capabilities
Some SharePoint customers are under the impression that the new AMSI integration with SharePoint Server would prevent files containing viruses from being uploaded to or downloaded from SharePoint Server. That isn't the purpose of the AMSI integration feature. Instead, it's to block malicious web requests from reaching SharePoint endpoints and compromising the server. For example, by exploiting a security vulnerability in a SharePoint endpoint before the official fix for the security vulnerability has been installed.

Updating the text of this article to better explain that the AMSI integration feature isn't a solution designed to prevent uploading or downloading files with viruses.