Update "Standard channel" chapter of this article
Based on this article: https://support.microsoft.com/en-us/office/demystifying-channels-c2356b56-2ff8-4391-a5ae-67963f9e19f6 we clearly state that the "General" channel can't be deleted. I would like to propose that we include this information here as well inside the "Standard channel" explanation so that this information is clearer to our users.