Update drives-used-for-sql-databases-are-running-out-of-free-space.md
The originally summary was factually incorrect and created a case of confusion for the customer. We don't need to be able to grow to twice its size, we need enough free space to autogrow TWICE. This is not a doubling of the space used. 

Edited so that the summary statement utilizes the exact same verbiage from the health analyzer rule "A drive should have enough free space for the largest database file on it to autogrow twice" and includes examples for clarity.