Fix error on roles with access to report. Add link for more info.
The usage report method lists global reader as a role required to view the report, however global reader role does not have access to the user data and so can't be used to run the report. removed from sentence and added link after for information on the roles that do have access.